+++ src/mdcraft/analysis/__init__.py
@@ -1,6 +1,6 @@
 """
-Simulation trajectory analysis
-==============================
+Simulation state data and trajectory analysis
+=============================================
 .. moduleauthor:: Benjamin Ye <GitHub: @bbye98>
 
 This module provides a variety of classes for analyzing simulation
@@ -10,5 +10,5 @@ trajectories.
 from . import (base, electrostatics, polymer, profile, reader, structure,
                thermodynamics, transport)
 
-__all__ = ["base", "electrostatics", "polymer", "profile", "reader", 
+__all__ = ["base", "electrostatics", "polymer", "profile", "reader",
            "structure", "thermodynamics", "transport"]
