+++ src/mdcraft/analysis/electrostatics.py
@@ -1,6 +1,6 @@
 """
-Electrostatics
-==============
+Electrostatic properties
+========================
 .. moduleauthor:: Benjamin Ye <GitHub: @bbye98>
 
 This module contains classes to quantify electrostatic properties, such
@@ -140,7 +140,7 @@ class DipoleMoment(DynamicAnalysisBase):
         .. note::
 
            If :code:`neutralize=True` or :code:`unwrap=True`, all atoms
-           of any particular molecule must belong to the same atom 
+           of any particular molecule must belong to the same atom
            group.
 
     charges : array-like, `openmm.unit.Quantity`, or \
