### 2024-07-06 ed2781d
Allow Hash objects to be put in NumPy arrays
