### 2024-07-11 #6957 @srmnitc
> Just to confirm, you'd be fine with me @{username} directly?

I realise I said 'with' instead of 'without'; sorry! I edited my comment above now. Please suggest only without @, and I will contact them. Sorry for the confusion, thanks again!
