### 2024-07-11 #6957 @pw0908
@srmnitc Oh, we need to update this. We have added examples as part of the quick start page: https://mdcraft.readthedocs.io/en/latest/notebooks/getting_started/quick_start.html

We were going to populate the User guide with the content from this repository, which is for a separate paper: https://github.com/bbye98/gcme

The API docs are very rigorous however.
