### 2024-07-16 63692f0
Fix formula for solvation potential
