### 2024-07-19 #7013 @editorialbot
[REVIEW]: MDCraft: A Python assistant for performing and analyzing molecular dynamics simulations of soft matter systems
**Submitting author:** @pw0908 (<a href="http://orcid.org/0000-0001-8628-6561">Pierre Walker</a>)
**Repository:** https://github.com/bbye98/mdcraft
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.0
