### 2024-07-19 bbye98/mdcraft#3 @srmnitc
CI tests failing
This is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/7013

Tests seem to be failing with `ModuleNotFoundError: No module named 'dynasor'`
