+++ lib/openmm-ic-plugin/CMakeLists.txt
@@ -1,0 +1,125 @@
+#-------------------------------------------------------------------------------
+# OpenMM Image Charge Plugin
+#
+# Creates the OpenMM image charge plugin library.
+#
+# Windows:
+#   OpenMMIC.dll
+#   OpenMMIC.lib
+# Unix:
+#   libOpenMMIC.so
+#-------------------------------------------------------------------------------
+
+CMAKE_MINIMUM_REQUIRED(VERSION 3.6)
+PROJECT(OpenMMICPlugin)
+CMAKE_POLICY(SET CMP0146 OLD)
+
+# We need to know where OpenMM is installed so we can access the headers and libraries.
+
+SET(OPENMM_DIR "/usr/local/openmm" CACHE PATH "Where OpenMM is installed")
+INCLUDE_DIRECTORIES("${OPENMM_DIR}/include")
+LINK_DIRECTORIES("${OPENMM_DIR}/lib" "${OPENMM_DIR}/lib/plugins")
+
+# Specify the C++ version we are building for.
+
+SET(CMAKE_CXX_STANDARD 11)
+
+# Set flags for linking on macOS.
+
+IF(APPLE)
+    SET (CMAKE_INSTALL_NAME_DIR "@rpath")
+    SET(EXTRA_COMPILE_FLAGS "-msse2 -stdlib=libc++")
+ENDIF(APPLE)
+
+# Select where to install.
+
+IF(${CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT})
+    IF(WIN32)
+        SET(CMAKE_INSTALL_PREFIX "$ENV{ProgramFiles}/OpenMM" CACHE PATH "Where to install the plugin" FORCE)
+    ELSE(WIN32)
+        SET(CMAKE_INSTALL_PREFIX "/usr/local/openmm" CACHE PATH "Where to install the plugin" FORCE)
+    ENDIF(WIN32)
+ENDIF(${CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT})
+
+# The source is organized into subdirectories, but we handle them all from
+# this CMakeLists file rather than letting CMake visit them as SUBDIRS.
+
+SET(IC_PLUGIN_SOURCE_SUBDIRS openmmapi serialization)
+
+# Set the library name.
+
+SET(IC_LIBRARY_NAME OpenMMIC)
+SET(SHARED_IC_TARGET ${IC_LIBRARY_NAME})
+
+# These are all the places to search for header files which are to be part of the API.
+
+SET(API_INCLUDE_DIRS "openmmapi/include" "openmmapi/include/openmm")
+
+# Locate header files.
+
+SET(API_INCLUDE_FILES)
+FOREACH(dir ${API_INCLUDE_DIRS})
+    FILE(GLOB fullpaths ${dir}/*.h)
+    SET(API_INCLUDE_FILES ${API_INCLUDE_FILES} ${fullpaths})
+ENDFOREACH(dir)
+
+# Collect source files.
+
+SET(SOURCE_FILES) # empty
+SET(SOURCE_INCLUDE_FILES)
+FOREACH(subdir ${IC_PLUGIN_SOURCE_SUBDIRS})
+    FILE(GLOB src_files  ${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/src/*.cpp)
+    FILE(GLOB incl_files ${CMAKE_CURRENT_SOURCE_DIR}/${subdir}/src/*.h)
[…]
