### 2024-07-19 29df535
Use Conda for installing Python packages in CI
