+++ .github/workflows/ci.yml
@@ -21,9 +21,7 @@ jobs:
           channels: conda-forge
           python-version: ${{ matrix.python-version }}
       - name: install-dependencies
-        run: |
-          python3 -m pip install -r requirements.txt
-          conda install openmm
+        run: conda env update --file environment.yml
       - name: lint
         run: ruff check --target-version=py39 .
         continue-on-error: true
