### 2024-07-19 f7f78df
Use Conda for installing Python packages in CI
