+++ .github/workflows/ci.yml
@@ -21,7 +21,9 @@ jobs:
           channels: conda-forge
           python-version: ${{ matrix.python-version }}
       - name: install-dependencies
-        run: conda env update --file environment.yml
+        run: |
+          conda activate test
+          conda env update --file environment.yml --verbose
       - name: lint
         run: ruff check --target-version=py39 .
         continue-on-error: true
