### 2024-07-19 2b781db
Use Conda for installing Python packages in CI
