### 2024-07-19 1f30f56
Use Conda for installing Python packages in CI
