+++ environment.yml
@@ -25,7 +25,7 @@ dependencies:
   # TEST DEPENDENCIES
   - ase
   - coverage
-  - dynasor
+  - dynasor >=2
   - mdanalysistests
   - pytest
   - ruff
