### 2024-07-19 2ff7dc2
Minor Ruff-recommended fixes
