### 2024-07-19 bbye98/mdcraft#3 @bbye98
Turns out `dynasor==2.0.0` requires `importlib-metadata<5`, but the newest `sphinx==7.4.6` requires `importlib-metadata>=6`. Fixed this issue by forcing `dynasor>=2` in the `requirements.txt` and `environment.yml` files.
