### 2024-07-19 d7fd26d
Minor Ruff-recommended fixes
