### 2024-07-24 605567e
Add preliminary multithreading support for LAMMPS dump reader
