+++ src/mdcraft/analysis/reader.py
@@ -6,9 +6,12 @@ Topology and trajectory readers
 This module contains topology and trajectory readers for use with MDAnalysis.
 """
 
+import concurrent.futures
+import os
 from typing import Union
 import warnings
 
+import psutil
 from MDAnalysis.coordinates.base import ReaderBase
 from MDAnalysis.lib.distances import transform_StoR
 from MDAnalysis.lib.util import anyopen, store_init_arguments
@@ -20,21 +23,21 @@ class LAMMPSDumpTrajectoryReader(ReaderBase):
 
     r"""
     LAMMPS dump trajectory reader.
-    
+
     Parameters
     ----------
     filename : `str`
         Filename of the LAMMPS dump file.
 
     conventions : `str` or `tuple`, optional
-        Coordinate conventions for the :math:`x`, :math:`y`, and 
+        Coordinate conventions for the :math:`x`, :math:`y`, and
         :math:`z`-positions. If a `str` is provided, the same value is
         used for all axes. Determined automatically if not provided.
 
         .. container::
 
            **Valid values:**
-           
+
            * :code:`None` to ignore an axis (when coordinate information
              is unavailable in that axis).
            * :code:`""` for unscaled and wrapped coordinates.
@@ -46,27 +49,27 @@ class LAMMPSDumpTrajectoryReader(ReaderBase):
         Determines whether atom positions are unwrapped.
 
     extras : `list`, optional
-        Extra per-atom information to be read from the dump file. 
-        Topology attributes, like type, mass, and charge, are not 
+        Extra per-atom information to be read from the dump file.
+        Topology attributes, like type, mass, and charge, are not
         expected to change and are not supported.
 
         .. note::
 
-           For named LAMMPS vector attributes, the stored array will 
+           For named LAMMPS vector attributes, the stored array will
            always be arrays of shape :math:`(N,\,3)`, where :math:`N` is
            the number of atoms. If data for an axis is not found in the
-           dump file, the corresponding column in the array will be 
+           dump file, the corresponding column in the array will be
            filled with zeros.
 
            For custom LAMMPS vector attributes, the stored array will
            have as many columns as there are instances of the attribute
-           in the dump file. For example, if :code:`"c_custom"` was 
-           provided to `extras` and the dump file contains the 
+           in the dump file. For example, if :code:`"c_custom"` was
+           provided to `extras` and the dump file contains the
            instances :code:`c_custom[2]`, :code:`c_custom[4]`, and
            :code:`c_custom[1]`, in that order, the array will have shape
            :math:`(N,\,3)`. The instances are always sorted, such that
-           the columns in the example array correspond to 
-           :code:`c_custom[1]`, :code:`c_custom[2]`, and 
+           the columns in the example array correspond to
+           :code:`c_custom[1]`, :code:`c_custom[2]`, and
            :code:`c_custom[4]`, respectively.
 
         **Valid values:**
@@ -75,7 +78,7 @@ class LAMMPSDumpTrajectoryReader(ReaderBase):
         Keyword (`Reader.data` key)         LAMMPS dump attribute(s)                             Type
         ==================================  ===================================================  ======================
         :code:`"dipole_moments"`            (:code:`mux`, :code:`muy`, :code:`muz`)              `numpy.ndarray[float]`
-        :code:`"dipole_moment_magnitudes"`  :code:`mu`                                           `float`     
[…]
