+++ paper/paper.bib
@@ -1,17 +1,33 @@
-@article{hautman_molecular_1989,
-	title = {Molecular dynamics simulation of water beween two ideal classical metal walls},
-	volume = {91},
-	issn = {0021-9606, 1089-7690},
-	url = {http://aip.scitation.org/doi/10.1063/1.457481},
-	doi = {10.1063/1.457481},
-	pages = {467--472},
-	number = {1},
-	journaltitle = {The Journal of Chemical Physics},
-	shortjournal = {J. Chem. Phys.},
-	author = {Hautman, J. and Halley, J. W. and Rhee, Y.‐J.},
-	urldate = {2020-12-03},
-	date = {1989-07},
-	langid = {english},
+
+@article{ballenegger_simulations_2009,
+	title = {Simulations of non-neutral slab systems with long-range electrostatic interactions in two-dimensional periodic boundary conditions},
+	volume = {131},
+	issn = {00219606},
+	url = {http://scitation.aip.org/content/aip/journal/jcp/131/9/10.1063/1.3216473},
+	doi = {10.1063/1.3216473},
+	language = {en},
+	number = {9},
+	urldate = {2022-03-17},
+	journal = {J. Chem. Phys.},
+	author = {Ballenegger, V. and Arnold, A. and Cerdà, J. J.},
+	year = {2009},
+	pages = {094107}
+}
+
+@article{fong_onsager_2020,
+	title = {Onsager {Transport} {Coefficients} and {Transference} {Numbers} in {Polyelectrolyte} {Solutions} and {Polymerized} {Ionic} {Liquids}},
+	volume = {53},
+	issn = {0024-9297, 1520-5835},
+	url = {https://pubs.acs.org/doi/10.1021/acs.macromol.0c02001},
+	doi = {10.1021/acs.macromol.0c02001},
+	language = {en},
+	number = {21},
+	urldate = {2022-01-12},
+	journal = {Macromolecules},
+	author = {Fong, Kara D. and Self, Julian and McCloskey, Bryan D. and Persson, Kristin A.},
+	month = nov,
+	year = {2020},
+	pages = {9503--9512}
 }
 
 @article{yeh_ewald_1999,
@@ -20,64 +36,48 @@
 	issn = {0021-9606, 1089-7690},
 	url = {http://aip.scitation.org/doi/10.1063/1.479595},
 	doi = {10.1063/1.479595},
-	pages = {3155--3162},
+	language = {en},
 	number = {7},
-	journaltitle = {The Journal of Chemical Physics},
-	shortjournal = {J. Chem. Phys.},
-	author = {Yeh, In-Chul and Berkowitz, Max L.},
 	urldate = {2021-09-01},
-	date = {1999-08-15},
-	langid = {english},
+	journal = {J. Chem. Phys.},
+	author = {Yeh, In-Chul and Berkowitz, Max L.},
+	month = aug,
+	year = {1999},
+	pages = {3155--3162}
 }
 
-@article{ballenegger_simulations_2009,
-	title = {Simulations of non-neutral slab systems with long-range electrostatic interactions in two-dimensional periodic boundary conditions},
-	volume = {131},
-	issn = {00219606},
-	url = {http://scitation.aip.org/content/aip/journal/jcp/131/9/10.1063/1.3216473},
-	doi = {10.1063/1.3216473},
-	pages = {094107},
-	number = {9},
-	journaltitle = {The Journal of Chemical Physics},
-	shortjournal = {J. Chem. Phys.},
-	author = {Ballenegger, V. and Arnold, A. and Cerdà, J. J.},
-	urldate = {2022-03-17},
-	date = {2009},
-	langid = {english},
-}
-
-@article{fong_onsager_2020,
[…]
