### 2024-07-24 b94c38f
Fix GCMe reference in paper.md
