+++ paper/paper.bib
@@ -212,3 +212,14 @@
 	year = {2024},
 	pages = {acs.jctc.4c00603}
 }
+
+@misc{lee_molecular_2024,
+	title = {Molecular {Control} via {Dynamic} {Bonding} {Enables} {Material} {Responsiveness} in {Additively} {Manufactured} {Metallo}-{Polyelectrolytes}},
+	copyright = {https://creativecommons.org/licenses/by/4.0/},
+	url = {https://www.researchsquare.com/article/rs-3643582/v1},
+	doi = {10.21203/rs.3.rs-3643582/v1},
+	urldate = {2024-07-24},
+	author = {Lee, Seola and Walker, Pierre and Velling, Seneca and Chen, Amylynn and Taylor, Zane and Fiori, Cyrus and Gandhi, Vatsa and Wang, Zhen-Gang and Greer, Julia},
+	month = jan,
+	year = {2024}
+}
