### 2024-07-24 #7013 @bbye98
I've updated the paper bibliography for a paper authored by me that recently got published. The only change is replacing the arXiv DOI with the J. Chem. Theory Comput. DOI.
