### 2024-07-25 #7013 @srmnitc
> @editorialbot generate pdf
>
> I've updated the paper bibliography for a paper authored by me that recently got published. The only change is replacing the arXiv DOI with the J. Chem. Theory Comput. DOI.

Thank you for the change!
