### 2024-07-25 ef05463
Optimize serial and parallel reading algorithms
