### 2024-07-29 bbye98/mdcraft#4 @raynol-dsouza
"latex not found" error on running the plot submodule code snippet in Quick Start
Hello,

I am one of the reviewers for your JOSS submission [https://github.com/openjournals/joss-reviews/issues/7013](url). I installed the package **on a fresh conda environment**, following your instructions [here](https://mdcraft.readthedocs.io/en/latest/notebooks/getting_started/installation.html). When running the code snippet for the `plot` submodule, I run into the following error:

`RuntimeError: Failed to process string with tex because latex could not be found`

The error could simply be due to setting `"text.usetex": True` on line 44 of `src/mdcraft/plot/rcparams.py` while not including any latex-rendering dependency in the conda recipe. Could you please look into this?

Thank you!
