### 2024-07-29 #7013 @raynol-dsouza
Dear all,

I find that the code for this package is well written and documented, with clear instructions for its installation. The tests seem comprehensive enough to offer thorough coverage of the core functionality. There is no doubt a substantial scholarly effort here, as evident from the source code.

My concern, however, is justifying the "extension to existing functionality" (under the checkpoint **Functionality**) from just a single example. The accompanying paper states numerous use cases for the package . I believe it is necessary to have at least a couple of more examples in well documented notebooks (for example, using the package to run and post-process a lammps simulation).

@srmnitc if this is not a necessary criteria, I can continue with my checklist.
