### 2024-07-29 4b333c1
Make LaTeX optional
