### 2024-07-29 bbye98/mdcraft#4 @bbye98
Hi! I've went ahead and made the `text.usetex` parameter use `True` only when LaTeX is found in commit 4b333c1. This should hopefully resolve the issue.
