### 2024-07-29 #7013 @bbye98
> @pw0908 @bbye98 I ran into a (maybe rather trivial) issue that I have raised [here](https://github.com/bbye98/mdcraft/issues/4).

I should have resolved this issue with [this push](https://github.com/bbye98/mdcraft/issues/4#issuecomment-2256606292). Please let me know!
