### 2024-07-29 #7013 @bbye98
@raynol-dsouza A number of the extensions we mention are used in our most recent paper, for which the source code is available [here](https://github.com/bbye98/gcme).

The OpenMM and LAMMPS input scripts in `/benchmark` use the custom GCMe pair potential in OpenMM, our implementation of the slab correction in OpenMM, and our implementation of the image charges in both OpenMM and LAMMPS.

The Jupyter Notebook `analysis_gcme.ipynb` showcases some of the analysis modules through the evaluation of the number density/charge density/electrostatic potential density and radial distribution function. I also have a radius of gyration example that I am happy to add.

We do plan on adding (and have already started working on) a user guide in the near future showing the entire process from setting up the simulation to analyzing the data (essentially just piecing together `nvt_polyanion_counterion_solvent.py` and `analysis_gcme.ipynb`). However, as I have my Ph.D. defense in less than two weeks, I do not currently have the bandwidth to finish drafting the user guide.
