### 2024-07-29 #7013 @pw0908
@bbye98 @raynol-dsouza I'll handle porting over the code to mdcraft. Alternatively, we could just explicitly reference the repository on the MDCraft website? Would that be sufficient @raynol-dsouza? The only thing is that those examples might be too dense; splitting them up will make it easier to follow.
