+++ docs/source/conf.py
@@ -52,6 +52,7 @@ intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/reference/", None)
 }
+myst_enable_extensions = ["dollarmath"]
 myst_heading_anchors = 6
 napoleon_numpy_docstring = True
 nb_execution_timeout = -1
