### 2024-07-29 #7013 @raynol-dsouza
> A number of the extensions we mention are used in our most recent paper, for which the source code is available [here](https://github.com/bbye98/gcme).

@bbye98 This is really, really nice!

While I personally would prefer having minimal working examples of these in an `/examples` directory, considering @bbye98's upcoming thesis defense and @pw0908's suggestion to explicitly reference the `gcme` repository, I defer to the better judgment of @srmnitc.

I am happy as long as there is a clear demonstration of a fair amount of functionality mentioned in the write-up!
