+++ .readthedocs.yaml
@@ -9,7 +9,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "mambaforge-24.3"
+    python: "mambaforge-latest"
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
