### 2024-07-29 60780d9
Add missing myst-nb dependency in environment.yml
