+++ environment.yml
@@ -18,6 +18,7 @@ dependencies:
   - openmm
   # DOCUMENTATION DEPENDENCIES
   - furo
+  - myst-nb
   - numpydoc
   - sphinx
   - sphinx-copybutton
