### 2024-07-29 772c362
Add missing myst-nb dependency in environment.yml
