+++ .readthedocs.yaml
@@ -20,3 +20,7 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 conda:
   environment: environment.yml
+python:
+  install:
+    - method: pip
+      path: .
