### 2024-07-29 bbye98/mdcraft#5 @aazocar
[General checks] Contribution and Authorship
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

> Does the full list of paper authors seem appropriate and complete?

Of course, there is no confusion regarding the authorship of B. Ye and P. Walker. However, there are some contradicting lists of authors. The metadata on pyproject.toml lists as authors B. Ye, P. Walker and A. Glisman, which is different from the list of the paper. Please clarify this and make sure all lists of authors are harmonized.
