### 2024-07-29 bbye98/mdcraft#6 @aazocar
[Functionality/Example usage] Module LAMMPS
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

It is stated on the paper (line 16) that the software extends the functionality of OpenMM and LAMMPS simulation packages, by introducing custom force fields, etc. This is not shown in any examples or usage guide specifically for LAMMPS. This is one of the core features mentioned in the paper and for a user to verify the functionality there should be an example.
