### 2024-07-29 bbye98/mdcraft#7 @aazocar
[Functionality/Example usage] Module OpenMM
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

Regarding this module, there is good API documentation but potential new users of the software would benefit from having more extensive examples for the openMM system extensions and tools https://mdcraft.readthedocs.io/en/latest/api/mdcraft.openmm.system.html#module-mdcraft.openmm.system

Specifically, those mentioned in the paper: add_image_charges and add_slab_correction

Another related aspect, is the Gaussian core model with smeared electrostatics (GCMe) mentioned on the paper, but it is not clear where this can be found on the API docs.
