### 2024-07-29 bbye98/mdcraft#8 @aazocar
[Functionality/Example usage] Algorithm and analysis modules
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

`algorithm` and `analysis` modules would also greatly benefit from example use cases.

Also, from the paper, most of the mentioned features correspond to the `analysis` module, it is not clear from the text what is covered in the `algorithm` module.
