### 2024-07-29 bbye98/mdcraft#9 @aazocar
[Performance] Performance claims of the software
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

In the paper the following claims are mentioned, some clarification / data or additional references would be good.

* Line 18: "..,  introducing custom force fields, such as the **efficient** and intuitive ... GCMe"
* Line 24: "...  trajectories stored in **memory-efficient** formats."
* Line 50: "... **performance optimizations** that are not typically available in other simulation packages."
