### 2024-07-29 bbye98/mdcraft#10 @aazocar
[Documentation] Statement of need
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

A statement of need is missing on the repo or the documentation page. Only a brief description is available.
