### 2024-07-29 bbye98/mdcraft#11 @aazocar
[Documentation] Automated tests
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

There are automated testing, and it is running on the CI. However, I would suggest to add a small section on how to run the tests locally in the documentation.
