### 2024-07-29 bbye98/mdcraft#12 @aazocar
[Software paper] Statement of need
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

In the general, the statement of need focuses more on reasserting and describing the features and modules of MDCraft, which is already done in the Summary section.

Furthermore, one of the main necessities described is to support on-the-fly analysis (Line 47); but there are no examples or further description of how MDCraft addresses this. Some sentences to clarify this point would be great.
