### 2024-07-29 #7013 @bbye98
@raynol-dsouza @pw0908

As a sign of good faith and to preview the completed guide, I have pushed the work-in-progress user guide for the GCMe project: https://mdcraft.readthedocs.io/en/latest/notebooks/user_guide/gcme_polyelectrolyte_edlc.html

When completed, it will have a full, working OpenMM input script and multiple analysis examples at the end.
