### 2024-07-29 bbye98/mdcraft#13 @aazocar
[Software paper] State of the field
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7013

> Do the authors describe how this software compares to other commonly-used packages?

I have checked this point, but it could be further strengthened on the text. The comparison to MDAnalysis and MDTraj could be more detailed. Additionally, maybe you could take a look and see if these other packages are relevant:
[freud](https://freud.readthedocs.io/en/latest/)
[pytraj](https://amber-md.github.io/pytraj/latest/analysis)
[pysimm](https://pysimm.org/)
