### 2024-07-29 #7013 @aazocar
@bbye98 @pw0908 @srmnitc
MDCraft is a very well "crafted" package :) There is good API documentation, robust testing and I was able to install and run the given example. I can also see there is considerable scholarly effort and added value.
I have now finished my initial review and I have opened some issues on the repo. Closing these issues would allow me to check the rest of the items in my review checklist.
In general, I would like to see more examples and support for the functionality claims stated on the paper. I also support the issue raised by @raynol-dsouza and having a User Guide, even if on the shorter side, will address most of my suggestions. I do not have extensive experience on soft matter systems, but for an interested potential user, like me, more examples would be a great help.
