+++ docs/source/conf.py
@@ -33,6 +33,7 @@ extensions = [
     "sphinx.ext.duration",
     "sphinx.ext.githubpages",
     "sphinx.ext.intersphinx",
+    "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
     "sphinx.ext.viewcode"
 ]
@@ -52,7 +53,7 @@ intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/reference/", None)
 }
-myst_enable_extensions = ["dollarmath"]
+myst_enable_extensions = ["amsmath", "dollarmath"]
 myst_heading_anchors = 6
 napoleon_numpy_docstring = True
 nb_execution_timeout = -1
