### 2024-07-29 bbye98/mdcraft#5 @pw0908
The short answer here is that, originally, we were going to implement an analysis suit that Alec had implemented. He graduated a few weeks ago and we didn't have the time to port it over. However, his discussions were very helpful in the design of this package and beta testing so we felt it fair to include him on the pyproject.toml. However, since he won't be a maintainer of MDCraft, we didn't feel it was appropriate to include him on the paper.

We would feel more comfortable keeping the authorships as they stand. The only alternative would be to remove Alec from the pyproject.toml. Let us know what you think with this context in mind.
