### 2024-07-29 bbye98/mdcraft#6 @pw0908
This is poor wording on our part. The features listed are only made available to OpenMM simulations. As for LAMMPS, only the analysis submodules are compatible with it. As such, any OpenMM examples we give will also be applicable to simulations run using LAMMPS. I'll re-word the paper accordingly.

The only LAMMPS-specific feature we provide is our own LAMMPSDUMP trajectory reader. This is currently being used in a rather recent project and, as it is such a niche feature, we didn't feel it appropriate to bring up in the main paper. I do think an example use of this specific reader should be given and will do so.
