### 2024-07-29 bbye98/mdcraft#7 @pw0908
A few updates on this issue:
1. @bbye98 Has added an extensive example going over the GCMe forcefield within the [docs](https://mdcraft.readthedocs.io/en/latest/notebooks/user_guide/gcme_polyelectrolyte_edlc.html).
2. As mentioned in an earlier comment in the paper issue, there is an extensive set of examples given in [this repository](https://github.com/bbye98/gcme/tree/main) using mdcraft. This is something the other reviewer has mentioned and, although we are discussion the option of simply referring to this repository within mdcraft, we can move some of these examples over if you believe it is necessary.
