### 2024-07-29 bbye98/mdcraft#8 @pw0908
For each of these:
1. `analysis`: We definitely should and will include examples for each of these. We'll probably include them within the API docs rather than standalone examples as these tools themselves are not inherently complicated.
2. `algorithm`: This module is intended to be more low-level compared to `analysis` as it provides functions used within the latter module. I will certainly clarify this distinction in the text. I am less certain how helpful examples will be here as, ideally, only advanced users should be looking at these functions and they are sufficiently commented to be easy to follow. Let us know what you think.
