### 2024-07-29 bbye98/mdcraft#9 @pw0908
The benchmarks of these implementations have been provided within @bbye98's [recent publication](https://doi.org/10.1021/acs.jctc.4c00603) Table 1. We can mention this manuscript within the repository. Additionally, the benchmark codes have also been provided in the [gcme repository](https://github.com/bbye98/gcme/tree/main/benchmark). As we may do with the examples, we could move these benchmarks into mdcraft.

Would a simple reference to the published text be sufficient? Or would you prefer that we provide all the codes explicitly within the repository?
