### 2024-07-29 bbye98/mdcraft#12 @pw0908
Addressing your comments:
1. While we do re-assert the features provided by mdcraft, we do highlight how these features fill gaps currently in the field. Firstly, in the case of OpenMM, we provide support throughout the simulation process, making it more accessible for novice users. Our analysis modules, in contrast to the ones existing in the field, provide more flexibility without requiring users to look at the low-level code. I suppose we could provide some context for the need for GCMe, image charge and slab corrections; however, this was the topic of @bbye98's other paper, running into the risk of plagiarism. As such, I can't think of much else to add to the statement of need that would remain MDCraft-specific.
2. Ill clarify this statement in the paper.
