### 2024-07-29 bbye98/mdcraft#13 @pw0908
I'll add a few sentences on how MDcraft compares to MDAnalysis, MDTraj, freud and pytraj. Pysimm seems to no longer be supported?
