### 2024-07-29 1aac96c
Clarify usage of MDCraft with LAMMPS ([Functionality/Example usage] Module LAMMPS #6)
