+++ paper/paper.md
@@ -29,9 +29,9 @@ bibliography: paper.bib
 
 MDCraft is a comprehensive Python package designed to enhance research workflows involving molecular dynamics (MD) simulations. It streamlines the entire process—from setting up and executing simulations to analyzing trajectories using sophisticated algorithms and visualizing results—making computational chemistry more accessible to a broader audience. At its core, MDCraft comprises three principal components.
 
-First, the `lammps` and `openmm` modules provide user-friendly tools to initialize, optimize, and run simulations, enabling the exploration of various large soft matter systems across different timescales. These modules extend the functionality of the LAMMPS [@thompson_lammps_2022] and OpenMM [@eastman_openmm_2017] simulation packages by introducing custom force fields, such as the efficient and intuitive Gaussian core model with smeared electrostatics (GCMe) [@ye_gcme_2024]; incorporating advanced techniques like the slab correction [@yeh_ewald_1999;@ballenegger_simulations_2009] and the method of image charges [@hautman_molecular_1989] for charged systems with slab geometries; facilitating coarse-grained MD simulations by scaling physical values by the fundamental quantities (mass $m$, length $d$, energy $\epsilon$, and Boltzmann constant $k_\mathrm{B}T$); and offering feature-rich readers and writers for topologies and trajectories stored in memory-efficient formats.
+First, the `openmm` module provides user-friendly tools to initialize, optimize, and run simulations, enabling the exploration of various large soft matter systems across different timescales. This module extend the functionality of the OpenMM [@eastman_openmm_2017] simulation package by introducing custom force fields, such as the efficient and intuitive Gaussian core model with smeared electrostatics (GCMe) [@ye_gcme_2024]; incorporating advanced techniques like the slab correction [@yeh_ewald_1999;@ballenegger_simulations_2009] and the method of image charges [@hautman_molecular_1989] for charged systems with slab geometries; facilitating coarse-grained MD simulations by scaling physical values by the fundamental quantities (mass $m$, length $d$, energy $\epsilon$, and Boltzmann constant $k_\mathrm{B}T$); and offering feature-rich readers and writers for topologies and trajectories stored in memory-efficient formats.
 
-Second, the `algorithm` and `analysis` modules offer optimized serial and multithreaded algorithms and analysis classes for evaluating structural, thermodynamic, and dynamic properties using thermodynamic state and trajectory data. These properties include, but are not limited to, static and dynamic structure factors, density and potential profiles, end-to-end vector autocorrelation functions for polymers, and Onsager transport coefficients [@fong_onsager_2020].
+Second, the `algorithm` and `analysis` modules offer optimized serial and multithreaded algorithms and analysis classes for evaluating structural, thermodynamic, and dynamic properties using thermodynamic state and trajectory data. These properties include, but are not limited to, static and dynamic structure factors, density and potential profiles, end-to-end vector autocorrelation functions for polymers, and Onsager transport coefficients [@fong_onsager_2020]. These modules are not limited to OpenMM only and can be used with simulation run in other packages such as LAMMPS [@thompson_lammps_2022] and GROMACS [@abrahamGROMACSHighPerformance2015].
 
 Finally, the `fit` and `plot` modules simplify the post-processing and visualization of data, aiding in the creation of aesthetically pleasing figures for scientific publications. These modules consist of models for curve fitting and helper functions that interface seamlessly with the commonly used SciPy [@virtanen_scipy_2020] and Matplotlib [@hunter_matplotlib_2007] libraries.
 
