+++ paper/paper.bib
@@ -223,3 +223,21 @@
 	month = jan,
 	year = {2024}
 }
+
+@article{abrahamGROMACSHighPerformance2015,
+  title = {{{GROMACS}}: {{High}} Performance Molecular Simulations through Multi-Level Parallelism from Laptops to Supercomputers},
+  shorttitle = {{{GROMACS}}},
+  author = {Abraham, Mark James and Murtola, Teemu and Schulz, Roland and P{\'a}ll, Szil{\'a}rd and Smith, Jeremy C. and Hess, Berk and Lindahl, Erik},
+  year = {2015},
+  month = sep,
+  journal = {SoftwareX},
+  volume = {1--2},
+  pages = {19--25},
+  issn = {2352-7110},
+  doi = {10.1016/j.softx.2015.06.001},
+  urldate = {2023-05-10},
+  abstract = {GROMACS is one of the most widely used open-source and free software codes in chemistry, used primarily for dynamical simulations of biomolecules. It provides a rich set of calculation types, preparation and analysis tools. Several advanced techniques for free-energy calculations are supported. In version 5, it reaches new performance heights, through several new and enhanced parallelization algorithms. These work on every level; SIMD registers inside cores, multithreading, heterogeneous CPU--GPU acceleration, state-of-the-art 3D domain decomposition, and ensemble-level parallelization through built-in replica exchange and the separate Copernicus framework. The latest best-in-class compressed trajectory storage format is supported.},
+  langid = {english},
+  keywords = {Free energy,GPU,Molecular dynamics,SIMD},
+  file = {/Users/pierrewalker/Zotero/storage/SRQHIS9H/Abraham et al. - 2015 - GROMACS High performance molecular simulations th.pdf;/Users/pierrewalker/Zotero/storage/ZK47EBKN/S2352711015000059.html}
+}
