### 2024-07-30 2a72fb8
Add some simple examples for the analysis classes ([Functionality/Example usage] Algorithm and analysis modules #8)
