### 2024-07-30 c6eb132
Add guidance for tests ([Documentation] Automated tests #11)
