### 2024-07-30 531c46c
Add computational benchmarks ([Performance] Performance claims of the software #9)
