+++ README.md
@@ -34,6 +34,13 @@ bond/pair potentials, support for NetCDF trajectories, and much more.
 * [`plot`](https://github.com/bbye98/mdcraft/tree/main/src/mdcraft/plot):
 Settings and additional functionality for Matplotlib figures.
 
+### Benchmarks
+
+The novel forcefield provided by MDCraft, Gaussian Core Model with smeared electrostatics (GCMe), provides multiple benefits discussed in more detail within our recent [publication](https://doi.org/10.1021/acs.jctc.4c00603). Of note is the computational speed-up obtained from GCMe, especially when used in OpenMM:
+
+![benchmarks](/assets/benchmarks.png)
+
+The codes used to generate these benchmarks are provided in the associated [repository](https://github.com/bbye98/gcme).
 ## Getting started
 
 ### Prerequisites
