### 2024-07-30 9d67f86
Add reference for properties ([Software paper] References #14)
