+++ paper/paper.bib
@@ -239,5 +239,15 @@
   abstract = {GROMACS is one of the most widely used open-source and free software codes in chemistry, used primarily for dynamical simulations of biomolecules. It provides a rich set of calculation types, preparation and analysis tools. Several advanced techniques for free-energy calculations are supported. In version 5, it reaches new performance heights, through several new and enhanced parallelization algorithms. These work on every level; SIMD registers inside cores, multithreading, heterogeneous CPU--GPU acceleration, state-of-the-art 3D domain decomposition, and ensemble-level parallelization through built-in replica exchange and the separate Copernicus framework. The latest best-in-class compressed trajectory storage format is supported.},
   langid = {english},
   keywords = {Free energy,GPU,Molecular dynamics,SIMD},
-  file = {/Users/pierrewalker/Zotero/storage/SRQHIS9H/Abraham et al. - 2015 - GROMACS High performance molecular simulations th.pdf;/Users/pierrewalker/Zotero/storage/ZK47EBKN/S2352711015000059.html}
+}
+
+@book{rubinsteinPolymerPhysics2003,
+  title = {Polymer {{Physics}}},
+  author = {Rubinstein, Michael and Colby, Ralph H.},
+  year = {2003},
+  month = aug,
+  publisher = {Oxford University Press},
+  address = {Oxford, New York},
+  abstract = {Polymer Physics thoroughly details the fundamental concepts of polymer melts, solutions, and gels in terms of both static structure and dynamics. It goes beyond other introductory polymer texts, deriving the essential tools of the physical polymer chemist or engineer without skipping any steps.  The book is divided into four parts. Part One summarizes the necessary concepts of a first course on polymers and covers the conformations of single polymer chains. Part Two deals with the thermodynamics of polymer solutions and melts, including chain conformations in those states. Part Three applies the concepts of Part Two to the formation and properties of polymer networks. Part Four explains the essential aspects of how polymers move in both melt and solution states.  The text assumes a working knowledge of calculus, physics, and chemistry, but no prior knowledge of polymers. It is ideal for upper-level undergraduate and first-year graduate courses in Condensed Matter Physics, Soft Materials, and Polymers.  Features               Presents established results in an easily accessible way               Emphasizes physical insight rather than mathematical rigor               Provides detailed experimental sections at the end of each chapter               Includes more than 200 illustrations and 350 exercises              ,                Polymer Physics thoroughly details the fundamental concepts of polymer melts, solutions, and gels in terms of both static structure and dynamics. It goes beyond other introductory polymer texts, deriving the essential tools of the physical polymer chemist or engineer without skipping any steps.  The book is divided into four parts. Part One summarizes the necessary concepts of a first course on polymers and covers the conformations of single polymer chains. Part Two deals with the thermodynamics of polymer solutions and melts, including chain conformations in those states. Part Three applies the concepts of Part Two to the formation and properties of polymer networks. Part Four explains the essential aspects of how polymers move in both melt and solution states.  The text assumes a working knowledge of calculus, physics, and chemistry, but no prior knowledge of polymers. It is ideal for upper-level undergraduate and first-year graduate courses in Condensed Matter Physics, Soft Materials, and Polymers.  Features               Presents established results in an easily accessible way               Emphasizes physical insight rather than mathematical rigor               Provides detailed experimental sections at the end of each chapter               Includes more than 200 illustrations and 350 exercises},
+  isbn = {978-0-19-852059-7},
 }
