### 2024-07-30 #7013 @srmnitc
@raynol-dsouza and @aazocar thank you for your extensive reviews, and @bbye98 and @pw0908, thanks for getting started on the issues.

After going through the issues, a common point that both reviewers have pointed out is the lack of examples that can help validate the functional claims of the software. The [new example](https://mdcraft.readthedocs.io/en/latest/notebooks/user_guide/gcme_polyelectrolyte_edlc.html) is definitely helpful, so is adding a link to the other repository where more examples can be found. However, the solution here is to add more examples to the user guide section. These need not be fully fledged use cases. For example, using custom force fields with OpenMM is mentioned as a feature of the paper. Some small examples which illustrate how one can switch to a custom force field that MDCraft provides would indeed be extremely useful. In general, my suggestion would be to also envision a new user who might not really be interested in running the full workflow as you have published before, but rather wants to use one of the modular features that MDCraft provides.

Thanks again for all the work from both the authors and reviewers!
