### 2024-07-30 bbye98/mdcraft#5 @aazocar
In my opinion, it would be better to _not_ have an additional author listed on the pyproject.toml who is not on the paper, in order to have a consistent and clear list of authors.

Another option would be to add A. Glisman to a short Acknowledgment / Contributors section in the readme.

Maybe @srmnitc can weigh in on this?
