### 2024-07-30 bbye98/mdcraft#5 @srmnitc
I think it is better to be consistent between the different places where the authors are listed.

@pw0908 Just for clarification from my side. _'However, since he won't be a maintainer of MDCraft, we didn't feel it was appropriate to include him on the paper.'_ The JOSS author guidelines do not need someone to be maintainer on the project. The [authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) also mention 'active project direction and other forms of non-code contributions' as a possible criteria for authorship. That being said, it is up to the authors to decide on the authorships. If A. Glisman is indeed not an author of the paper, please keep it consistent in the repo, and maybe acknowledge them elsewhere as @aazocar points out.
