### 2024-07-30 bbye98/mdcraft#8 @aazocar
1. Including examples in the API docs addressed this point.
2. For `algorithm` module, I agree the examples here are not needed. The comment was intended to clarify the differences between both modules and highlighting the distinction within the text should be enough to close this issue.
