### 2024-07-30 bbye98/mdcraft#9 @aazocar
I would say that adding the benchmark section in the README is sufficient to support the performance statements  in lines 18 and 50 with regards to GCMe and `openmm` module.  I would say there is no need to move the benchmarks to the mdcraft repository.

One question would still be, what are the memory-efficient formats mentioned in line 24? Is this specifically for I/O? Would you say this is also addressed in the benchmark section?
