### 2024-07-30 6ed2562
Clarifying the distinction between analysis and algorithms ([Functionality/Example usage] Algorithm and analysis modules #8)
