### 2024-07-30 bbye98/mdcraft#9 @pw0908
I think there may have been a misunderstanding at this line: we can read and write memory-efficient formats (such as netcdf), but this is just using already-established file formats which we didn't implement. As such, I don't think it's on us to justify this claim.
