### 2024-07-30 dabff8b
Give example of memory-efficient format ([Performance] Performance claims of the software #9)
