### 2024-07-30 bbye98/mdcraft#5 @pw0908
One observation I will briefly make is that it isn't uncommon for the authors on the JOSS paper not to match the authors on the pyproject.toml. Here are a few examples where the JOSS authorship doesn't match the repository authorship:
[stimupy](https://github.com/computational-psychology/stimupy)
[atlite](https://github.com/PyPSA/atlite)
[pycm](https://github.com/sepandhaghighi/pycm)
[Gridap.jl](https://github.com/gridap/Gridap.jl)

These are the cases where there were more authors in the .toml file than the JOSS paper. More often than not, the inverse is usually true, with most cases only having the first author on the .toml, even if the latter authors contributed to the code. I suppose this behaviour is to be expected as open-source projects are inherently collaborative and who counts as an author can change even after the JOSS is published.

All this to say that there is precedent in JOSS for this to occur and that we would much prefer keeping the authorship as it stands. That being said, if this isn't enough to convince you, we'll just remove A. Glisman from our toml.
