### 2024-07-30 3f25c4e
Add SoN in docs ([Documentation] Statement of need #10)
