### 2024-07-30 a74a4d8
Clarify the purpose of the analysis class ([Software paper] Statement of need #12)
