### 2024-07-30 bbye98/mdcraft#5 @srmnitc
@pw0908 Thanks for bringing this to attention. In the list you provided, pycm and gridap.jl have matching list of authors for the version of the code reviewed in JOSS. atlite has a list of authors and contributors. That being said, I am sure you would be able to find more cases where there are differences.

There are four different places where authors are listed:

- pyproject.toml: B. Ye, P. Walker, A. Glisman
- commit history: B. Ye, P. Walker
- copyright statement in docs: B. Ye, P. Walker, A. Glisman
- paper: B. Ye, P. Walker, Z-G Wang

From this, there are two main things you could clarify that would shed some light on the question "Does the full list of paper authors seem appropriate and complete?"

1. Is A. Glisman an author of the paper? As I mentioned, one does not need to be a maintainer of the code to be a JOSS author. Since you mentioned they were involved in the design of the package, and they are listed on pyproject.toml and copyright statement, could you please confirm that A. Glisman indeed need not be an author of the paper?

2. Open-source projects are inherently collaborative, true, but A. Glisman does not have any commits in this repo. Keeping this in mind, my strong recommendation still is to harmonise the metadata and prevent four conflicting list of authors. I agree this is not a point that directly appears in the JOSS checklist, so the final decision would be up to you.

Thanks for the timely resolution of the review issues!
