### 2024-07-30 bbye98/mdcraft#5 @pw0908
Answering your questions:
1. A. Glisman was a beta tester of MDCraft (back when it was called MDHelper) and wasn't actually part of the design process. There was a chunk of code they wrote which was always intended to merge in formally but never got around to. As such, because they had been part of the MDCraft journey since the start and we had always intended to include a contribution from them, we felt a good compromise was acknowledging them in the pyproject.toml. That being said, @bbye98  I don't think Alec should be in the License? If anything, Z-G should be.
2. Fair point about harmonising the four different files... @bbye98 that does seem cleaner?
