### 2024-07-30 0472bf2
Add a bit more to the statement of need ([Software paper] Statement of need #12)
