+++ paper/paper.md
@@ -41,7 +41,7 @@ Together, these modules provide both novice and experienced MD simulation users 
 
 Although established MD analysis packages such as MDAnalysis [@michaudagrawal_mdanalysis_2011] and MDTraj [@mcgibbon_mdtraj_2015] have been around for a considerable time, they primarily focus on the post-simulation analysis. In contrast, MDCraft is designed to provide comprehensive support throughout the entire simulation process, from initialization to post-processing. 
 
-MDCraft is tightly integrated with OpenMM, a relatively new simulation toolkit that has seen a surge in popularity in recent years due to its class-leading performance and flexibility through support for custom intermolecular forces and integrators for equations of motion. Due to its age and design philosophy, OpenMM offers comparatively fewer choices of pair potentials and external forces, and no built-in analysis support. MDCraft fills this gap in two ways. First, the `openmm` module leverages the modularity of OpenMM to provide a suite of custom force fields, problem-solving tools, trajectory readers and writers, and utility functions for unit reduction, topology transformations, and performance optimizations that are not typically available in other simulation packages. Then, the classes in the `analysis` module enable computing common structural, thermodynamic, and dynamic properties using the topology, trajectory, and state data generated by OpenMM (or other simulation packages).
+MDCraft is tightly integrated with OpenMM, a relatively new simulation toolkit that has seen a surge in popularity in recent years due to its class-leading performance and flexibility through support for custom intermolecular forces and integrators for equations of motion. Due to its age and design philosophy, OpenMM offers comparatively fewer choices of pair potentials and external forces, and no built-in analysis support. MDCraft fills this gap in two ways. First, the `openmm` module leverages the modularity of OpenMM to provide a suite of custom force fields, problem-solving tools, trajectory readers and writers, and utility functions for unit reduction, topology transformations, and performance optimizations that are not typically available in other simulation packages. Of special significance is the support for the GCMe force field which, as demonstrated in a recent article[@ye_gcme_2024], provides substantial acceleration compared to other force fields while also remaining physically meaningful. Then, the classes in the `analysis` module enable computing common structural, thermodynamic, and dynamic properties using the topology, trajectory, and state data generated by OpenMM (or other simulation packages).
 
 The `analysis` module also stands out due to the remarkable flexibility it affords its end users. General users have unprecedented control over what aspects of the properties to calculate and which method to employ through a plethora of well-documented built-in options in each analysis class, without having to be concerned about the underlying implementations. More advanced users, on the other hand, have the option to work directly with the algorithms in the `algorithms` module for further customization. These analysis functions and classes have proven indispensable in several recent publications [@glisman_multivalent_2024;@mantha_adsorption_2024;@lee_molecular_2024].
 
