### 2024-07-31 2d67bc5
Add references ([Software paper] References #14)
