+++ paper/paper.bib
@@ -251,3 +251,89 @@
   abstract = {Polymer Physics thoroughly details the fundamental concepts of polymer melts, solutions, and gels in terms of both static structure and dynamics. It goes beyond other introductory polymer texts, deriving the essential tools of the physical polymer chemist or engineer without skipping any steps.  The book is divided into four parts. Part One summarizes the necessary concepts of a first course on polymers and covers the conformations of single polymer chains. Part Two deals with the thermodynamics of polymer solutions and melts, including chain conformations in those states. Part Three applies the concepts of Part Two to the formation and properties of polymer networks. Part Four explains the essential aspects of how polymers move in both melt and solution states.  The text assumes a working knowledge of calculus, physics, and chemistry, but no prior knowledge of polymers. It is ideal for upper-level undergraduate and first-year graduate courses in Condensed Matter Physics, Soft Materials, and Polymers.  Features               Presents established results in an easily accessible way               Emphasizes physical insight rather than mathematical rigor               Provides detailed experimental sections at the end of each chapter               Includes more than 200 illustrations and 350 exercises              ,                Polymer Physics thoroughly details the fundamental concepts of polymer melts, solutions, and gels in terms of both static structure and dynamics. It goes beyond other introductory polymer texts, deriving the essential tools of the physical polymer chemist or engineer without skipping any steps.  The book is divided into four parts. Part One summarizes the necessary concepts of a first course on polymers and covers the conformations of single polymer chains. Part Two deals with the thermodynamics of polymer solutions and melts, including chain conformations in those states. Part Three applies the concepts of Part Two to the formation and properties of polymer networks. Part Four explains the essential aspects of how polymers move in both melt and solution states.  The text assumes a working knowledge of calculus, physics, and chemistry, but no prior knowledge of polymers. It is ideal for upper-level undergraduate and first-year graduate courses in Condensed Matter Physics, Soft Materials, and Polymers.  Features               Presents established results in an easily accessible way               Emphasizes physical insight rather than mathematical rigor               Provides detailed experimental sections at the end of each chapter               Includes more than 200 illustrations and 350 exercises},
   isbn = {978-0-19-852059-7},
 }
+
+@article{freud2020,
+    title = {freud: A Software Suite for High Throughput
+             Analysis of Particle Simulation Data},
+    author = {Vyas Ramasubramani and
+              Bradley D. Dice and
+              Eric S. Harper and
+              Matthew P. Spellings and
+              Joshua A. Anderson and
+              Sharon C. Glotzer},
+    journal = {Computer Physics Communications},
+    volume = {254},
+    pages = {107275},
+    year = {2020},
+    issn = {0010-4655},
+    doi = {https://doi.org/10.1016/j.cpc.2020.107275},
+    url = {http://www.sciencedirect.com/science/article/pii/S0010465520300916},
+    keywords = {Simulation analysis, Molecular dynamics,
+                Monte Carlo, Computational materials science},
+}
+
+@article{roePTRAJCPPTRAJSoftware2013,
+  title = {{{PTRAJ}} and {{CPPTRAJ}}: {{Software}} for {{Processing}} and {{Analysis}} of {{Molecular Dynamics Trajectory Data}}},
+  shorttitle = {{{PTRAJ}} and {{CPPTRAJ}}},
+  author = {Roe, Daniel R. and Cheatham, Thomas E. III},
+  year = {2013},
+  month = jul,
+  journal = {Journal of Chemical Theory and Computation},
+  volume = {9},
+  number = {7},
+  pages = {3084--3095},
+  publisher = {American Chemical Society},
[…]
