### 2024-07-31 6c5958e
Remove A. Glisman ([General checks] Contribution and Authorship #5)
