+++ pyproject.toml
@@ -7,8 +7,7 @@ name = "mdcraft"
 version = "1.1.0"
 authors = [
   { name = "Benjamin B. Ye", email = "bye@caltech.edu" },
-  { name = "Pierre J. Walker", email = "pjwalker@caltech.edu" },
-  { name = "Alec G. Glisman", email = "aglisman@caltech.edu" }
+  { name = "Pierre J. Walker", email = "pjwalker@caltech.edu" }
 ]
 classifiers = [
   "Intended Audience :: Science/Research",
