### 2024-08-02 bbye98/mdcraft#5 @aazocar
> @pw0908 Thanks for addressing this issue. All the records are now consistent, @aazocar do you feel the issue is addressed, and we can close this?

Yes, from my side the issue has been clarified. I will close.
