### 2024-08-02 #7013 @aazocar
All the issues from my review have been closed now and I have completed all points from the review checklist. Thanks to @pw0908 and @bbye98 for addressing all aspects and making the changes.

A recommendation for the future would be to increase the examples on the User Guide, not necessarily in very long or dense examples, but rather small blocks that showcase specific functions. In my experience, the support of new and less experienced users helps to increase the re-usability of the software.

Another small suggestion would be to include a [CITATION.cff](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files) file in the repository to clearly state how others should cite the software.

Again, these are small recommendations and not blocking acceptance. Therefore, I recommend the publication of MDCraft in JOSS.

Thanks for the opportunity to review this software.
