### 2024-08-02 #7013 @bbye98
@aazocar @raynol-dsouza Thank you both so much for your time. I will definitely work to incorporate your recommendations and suggestions, and I will update the User Guide section in the upcoming month with a number of examples to further highlight the various functionalities of MDCraft. This is a project near and dear to my heart, so I expect to continue developing, contributing to, and maintaining for a long time. I am glad to hear that you both found the package useful.
