+++ src/mdcraft/analysis/thermodynamics.py
@@ -22,8 +22,8 @@ from .base import Hash
 if FOUND_OPENMM:
     from openmm import unit
 
-class ConstantVolumeHeatCapacity:
 
+class ConstantVolumeHeatCapacity:
     """
     A serial implementation to caluclate the constant-volume heat
     capacity :math:`C_V` for a canonical (:math:`NVT`) system.
@@ -111,30 +111,49 @@ class ConstantVolumeHeatCapacity:
 
     _COLUMNS = {
         "lammps": {
-            "energy": ["TotEng", "KinEng", "PotEng", "E_angle", "E_bond",
-                       "E_coul", "E_dihed", "E_impro", "E_long", "E_vdwl"],
-            "temperature": "Temp"
+            "energy": [
+                "TotEng",
+                "KinEng",
+                "PotEng",
+                "E_angle",
+                "E_bond",
+                "E_coul",
+                "E_dihed",
+                "E_impro",
+                "E_long",
+                "E_vdwl",
+            ],
+            "temperature": "Temp",
         },
         "openmm": {
-            "energy": ["Total Energy (kJ/mole)", "Kinetic Energy (kJ/mole)",
-                       "Potential Energy (kJ/mole)"],
-            "temperature": "Temperature (K)"
-        }
+            "energy": [
+                "Total Energy (kJ/mole)",
+                "Kinetic Energy (kJ/mole)",
+                "Potential Energy (kJ/mole)",
+            ],
+            "temperature": "Temperature (K)",
+        },
     }
 
     def __init__(
-            self, log_file: Union[str, Path] = None, log_format: str = None, *,
-            energies: Union[np.ndarray[float], "unit.Quantity", Q_] = None,
-            temperature: Union[float, "unit.Quantity", Q_] = None,
-            reduced: bool = False, sep: str = ",") -> None:
+        self,
+        log_file: Union[str, Path] = None,
+        log_format: str = None,
+        *,
+        energies: Union[np.ndarray[float], "unit.Quantity", Q_] = None,
+        temperature: Union[float, "unit.Quantity", Q_] = None,
+        reduced: bool = False,
+        sep: str = ",",
+    ) -> None:
 
         self.results = Hash(units={})
         self._reduced = reduced
 
         if energies:
             self.results.units["energies"] = ureg.kilojoule / ureg.mole
-            self.results.energies = strip_unit(energies,
-                                               self.results.units["energy"])[0]
+            self.results.energies = strip_unit(energies, self.results.units["energy"])[
+                0
+            ]
         elif log_file:
             self._file = log_file if isinstance(log_file, Path) else Path(log_file)
             with open(self._file, "r") as f:
@@ -154,7 +173,7 @@ class ConstantVolumeHeatCapacity:
             # Pre-process LAMMPS log files
             if self._format == "lammps":
                 if "minimize" in log:
-                    log = log[log.index("Minimization stats:"):]
+                    log = log[log.index("Minimization stats:") :]
                 log = log.split("\n")
                 valid = False
                 for i, line in enumerate(log):
@@ -163,22 +182,19 @@ class ConstantVolumeHeatCapacity:
                         valid = True
                         break
                 if not valid:
-                    emsg = ("No thermodynamic data found in log file "
-                            f"'{log_file}'.")
+                    emsg = "No thermodynamic data found in log file " f"'{log_file}'."
                     raise ValueError(emsg)
                 log = "\n".join(log)
-                log = log[:log.index("Loop time of ")]
-                kwargs = {"sep": "\s+"}
+                log = log[: log.index("Loop time of ")]
+                kwargs = {"sep": r"\s+"}
                 self.results.units["energies"] = ureg.kilocalorie / ureg.mole
-                self.results.units["heat_capacity"] \
-                    = ureg.kilocalorie / ureg.kelvin
[…]
