+++ src/mdcraft/__init__.py
@@ -13,9 +13,8 @@ FOUND_OPENMM = find_spec("openmm") is not None
 Q_ = Quantity
 ureg = UnitRegistry(auto_reduce_dimensions=True)
 
-from . import algorithm, analysis, fit, lammps, plot # noqa: E402
+from . import algorithm, analysis, fit, lammps, plot  # noqa: E402
 
-__all__ = ["FOUND_OPENMM", "VERSION", "algorithm", "analysis", "fit", "lammps",
-           "plot"]
+__all__ = ["FOUND_OPENMM", "VERSION", "algorithm", "analysis", "fit", "lammps", "plot"]
 if FOUND_OPENMM:
     __all__.append("openmm")

[…]
