### 2024-08-07 7fa4540
Add Python 3.9 support for LAMMPS dump reader
