### 2024-08-08 bbye98/mdcraft#16 @srmnitc
Changes to the paper
This is part of review over at https://github.com/openjournals/joss-reviews/issues/7013

Some final changes to the paper from my side. I have also opened a PR, feel free to merge that after verification or implement the changes yourself.

- line 35: 'aesthetically pleasing figures' is a rather subjective statement. If the figures conform to norms by publishers, please indicate that.
- line 45-46: 'not only improves upon the analysis tools provided' . I understand from this statement that MDcraft improves on the analysis tools provided the packages listed in the sentence before. These packages also have other functionality, so this statement may not be entirely accurate. If MDCraft improves on particular aspects, could you please clarify this in the paper; or change this sentence to something more neutral.
- line 50-51: needs citation to strengthen the argument
- line 62-65: has a number of words that might sound superflous to the reader. 'remarkable flexibility', and 'unprecedent control' are also claims that are rather hard to verify. Could you please consider rephrasing this sentence?
- citation for GCMe is missing pages. The citation file provided by the journal has 'null' in pages. Could you please recheck this.

Thanks a lot for all the work!
