### 2024-08-08 #7013 @srmnitc
@aazocar and @raynol-dsouza Once again, thanks a lot for taking the time to review the packages and for providing detailed feedback. Please consider signing up at https://reviewers.joss.theoj.org/join as reviewers if you have not done already :) Thank you for your efforts!

@bbye98 @pw0908 Thanks for implementing the changes. I have opened an issue and a PR, after addressing this we can move on with the rest of steps.
