### 2024-08-08 8a8365f
Address: Changes to the paper #16
