+++ paper/paper.md
@@ -33,17 +33,17 @@ First, the `openmm` module provides user-friendly tools to initialize, optimize,
 
 Second, the `algorithm` and `analysis` modules offer optimized serial and multithreaded algorithms and analysis classes for evaluating structural, thermodynamic, and dynamic properties using thermodynamic state and trajectory data. The analysis classes provide properties including, but not limited to, static and dynamic structure factors [@faberTheoryElectricalProperties1965;@ashcroftStructureBinaryLiquid1967;@rogMoldynProgramPackage2003], density and potential profiles, end-to-end vector autocorrelation functions for polymers, and Onsager transport coefficients [@rubinsteinPolymerPhysics2003;@fong_onsager_2020]. The algorithms provide the underlying tools used to perform analysis and are intended to be easily extensible by more-advanced users. These modules are not limited to OpenMM only and can be used with simulation run in other packages such as LAMMPS [@thompson_lammps_2022] and GROMACS [@abrahamGROMACSHighPerformance2015].
 
-Finally, the `fit` and `plot` modules simplify the post-processing and visualization of data, aiding in the creation of aesthetically pleasing figures for scientific publications. These modules consist of models for curve fitting and helper functions that interface seamlessly with the commonly used SciPy [@virtanen_scipy_2020] and Matplotlib [@hunter_matplotlib_2007] libraries.
+Finally, the `fit` and `plot` modules simplify the post-processing and visualization of data, aiding in the creation of figures suitable for scientific publications. These modules consist of models for curve fitting and helper functions that interface seamlessly with the commonly used SciPy [@virtanen_scipy_2020] and Matplotlib [@hunter_matplotlib_2007] libraries.
 
 Together, these modules provide both novice and experienced MD simulation users with a comprehensive set of tools necessary to conduct computer experiments ranging from simple to complex, all within a single, succinct package.
 
 # Statement of need
 
-Although established MD analysis packages such as MDAnalysis [@michaudagrawal_mdanalysis_2011], MDTraj [@mcgibbon_mdtraj_2015], freud [@freud2020] and pytraj [@roePTRAJCPPTRAJSoftware2013] have been around for a considerable time, they primarily focus on the post-simulation analysis. In contrast, MDCraft not only improves upon the analysis tools provided, but is designed to also provide comprehensive support throughout the entire simulation process, from initialization to post-processing. 
+Although established MD analysis packages such as MDAnalysis [@michaudagrawal_mdanalysis_2011], MDTraj [@mcgibbon_mdtraj_2015], freud [@freud2020] and pytraj [@roePTRAJCPPTRAJSoftware2013] have been around for a considerable time, they primarily focus on the post-simulation analysis. In contrast, MDCraft not only improves upon some of the analysis tools provided, but is designed to also provide comprehensive support throughout the entire simulation process, from initialization to post-processing. 
 
[…]
