### 2024-08-08 bbye98/mdcraft#16 @pw0908
Hi!

Answering your comments:
* Wording was changed to remove the subjective statement and highlight that the plots would be suitable for publication.
* We've made this sentence a bit more neutral.
* We've added a citation for this.
* We've slightly reworded this sentence, removing superflous-sounding words.
* There's not a lot we can do about this. The paper was only published a few days ago and doesn't have an issue number / page number yet...
