### 2024-08-12 #7013 @pw0908
@srmnitc I can confirm the authors, affiliations and ORCIDs are all correct. I have also created the automated .zenodo.json file with the same information as the JOSS paper. @bbye98 just passed his defense last friday; once I can get a hold of him again, I'll have him make the release and zenodo repos.
