### 2024-08-12 #7013 @pw0908
That was a bit of a struggle. But here are the version number and zenodo DOI:
* Version number: 1.2.0
* Zenodo DOI: 10.5281/zenodo.13308642
There was a small issue with zenodo and it made two copies of 1.2.0. The DOI above is the most up-to-date and, inherently, they are identical.
