### 2024-08-13 #7013 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
