### 2024-08-15 #7013 @Kevin-Mattheus-Moerman
@pw0908 as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. All seems in order, so I'll now proceed to accept this submission.
